Improve vcs-compare performance while maintaining compatability
This version of improved `vcs-compare.sh` should maintain the
functionality of the original script. However, the time required to
execute has been significantly reduced. For my machine using unix `time`
on a early 2011 MacBookPro (15 in. core i7 with 8 GB ram):

1. original (vcs-compare v1.0):

    ./segments/vcs_compare.sh  0.03s user 0.04s system 62% cpu 0.119 total

2. reduced-functionality for speed (PR #90):

    ./segments/vcs_compare.sh  0.02s user 0.03s system 68% cpu 0.076 total

3. this commit:

    ./segments/vcs_compare.sh  0.02s user 0.03s system 61% cpu 0.084 total

As can be seen, this version is not the fastest possible, but it may be
the fastest possible version capable of retaining functionality and
compatability.